+++ packages/create-svelte/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-svelte
 
+## 7.0.0
+### Major Changes
+
+
+- breaking: the `create-svelte` package has been replaced by the `sv` package ([#12850](https://github.com/sveltejs/kit/pull/12850))
+
 ## 6.4.0
 ### Minor Changes